pyproject.toml
[project]
name = "knowledge-hub"
version = "0.1.0"
description = "Slack-to-Notion knowledge base automation pipeline"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.41.0",
    "pydantic-settings>=2.13.0",
    "slack-sdk>=3.40.1",
    "httpx>=0.28.1",
    "trafilatura>=2.0.0",
    "youtube-transcript-api>=1.2.4",
    "pypdf>=6.7.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.0.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
